+++ features_pathrslite.go
@@ -1,0 +1,7 @@
+//go:build !libpathrs
+
+package main
+
+func pathrsVersionString() string {
+	return ""
+}